% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Odds.Ratio.Binary}
\alias{Odds.Ratio.Binary}
\title{Produce binary odds ratios}
\usage{
Odds.Ratio.Binary(fit, target, evidence.node, reference.level,
  evidence.level, target.ref, target.target)
}
\description{
Produce binary odds ratios
}
