% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse.cpdist}
\alias{parse.cpdist}
\title{Parse a cpdist query for a given mutilated network, target and evidence node + level}
\usage{
parse.cpdist(mutilated.network, target, node, level)
}
\arguments{
\item{mutilated.network}{a fit mutilated network}

\item{target}{the target to sample in respect to}

\item{node}{the evidence node}

\item{level}{the level of the evidence node}
}
\value{
return the probabilities of the target under evidence
}
\description{
Parse a cpdist query for a given mutilated network, target and evidence node + level
}
