% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Produce.OR}
\alias{Produce.OR}
\title{Produce Odds Ratios plots for a given bnlearn fit object, and configuration file.}
\usage{
Produce.OR(fit, config)
}
\arguments{
\item{fit}{the bnlearn fit object}

\item{config}{the loaded yaml configuration file}
}
\description{
Produce Odds Ratios plots for a given bnlearn fit object, and configuration file.
}
\examples{
config <- yaml::read_yaml('config.yml')
Produce.OR(fit, config)
}
